[TLX] Fix TMA test type mismatches and CoalesceAsyncCopy bulk copy crash
1. Fix tl.int16 → tl.float16 in TMA tests that use torch.float16
   input. The TMA verifier now requires exact element type matching.

2. Fix CoalesceAsyncCopy crash on bulk async copies: use dyn_cast
   instead of cast for RankedTensorType since bulk copies use a
   scalar !tt.ptr source, not a tensor of pointers.

+++ python/test/unit/language/test_tlx_tma.py
@@ -731,7 +731,7 @@ def test_prefetch_tensormap(device):
         tlx.prefetch(in_desc, tensormap=True)
         tlx.prefetch(out_desc, tensormap=True)
 
-        buffers = tlx.local_alloc((BLOCK_SIZE_M, BLOCK_SIZE_N), tl.float16, tl.constexpr(1))
+        buffers = tlx.local_alloc((BLOCK_SIZE_M, BLOCK_SIZE_N), tl.int16, tl.constexpr(1))
         buffer = tlx.local_view(buffers, 0)
         bars = tlx.alloc_barriers(tl.constexpr(1))
         bar = tlx.local_view(bars, 0)
@@ -792,7 +792,7 @@ def test_prefetch_tensormap(device):
         tlx.prefetch(desc_in, tensormap=True)
         tlx.prefetch(desc_out, tensormap=True)
 
-        buffers = tlx.local_alloc((BLOCK_SIZE_M, BLOCK_SIZE_N), tl.float16, tl.constexpr(1))
+        buffers = tlx.local_alloc((BLOCK_SIZE_M, BLOCK_SIZE_N), tl.int16, tl.constexpr(1))
         buffer = tlx.local_view(buffers, 0)
         bars = tlx.alloc_barriers(tl.constexpr(1))
         bar = tlx.local_view(bars, 0)
